@@ -8,7 +8,7 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.2.2.RELEASE</version>
+    <version>2.6.2</version>
   </parent>
 
   <!-- (2) <groupId/> -->
@@ -19,7 +19,7 @@
   <artifactId>hello</artifactId>
 
   <!-- (4) <version/> -->
-  <version>1.0.0</version>
+  <version>1.1.0</version>
 
   <!-- (5) <packaging/> -->
   <!-- (6) <name/> -->
@@ -43,7 +43,7 @@
   <!-- (21) <properties/> -->
 
   <properties>
-    <java.version>11</java.version>
+    <java.version>17</java.version>
     <mainClass>edu.ucsb.cs156.spring.hello.Application</mainClass>
   </properties>
 
@@ -88,7 +88,7 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-jar-plugin</artifactId>
-        <version>3.1.2</version>
+        <version>3.2.0</version>
         <configuration>
           <archive>
             <manifest>
@@ -103,7 +103,7 @@
       <plugin>
         <groupId>org.jacoco</groupId>
         <artifactId>jacoco-maven-plugin</artifactId>
-        <version>0.8.5</version>
+        <version>0.8.7</version>
         <executions>
           <execution>
             <goals>
@@ -123,7 +123,7 @@
       <plugin>
         <groupId>org.pitest</groupId>
         <artifactId>pitest-maven</artifactId>
-        <version>1.5.2</version>
+        <version>1.7.3</version>
         <configuration>
           <targetClasses>
             <param>edu.*</param>
